Code Refactoring
* Replaced Guava Strings#isNullOrEmpty() with StringUtils#isBlank()
* Formatting
* Extracted PropertiesForm to a centralized place
