+++ com.osgifx.console.util/src/main/java/com/osgifx/console/util/fx/Fx.java
@@ -233,4 +233,5 @@ public final class Fx {
         }
         return button;
     }
+
 }
