+++ com.osgifx.console.util/src/main/resources/kv.png
